Update dependencies, adjust to pyo3 0.22. Respect virtualenv (kind of).
I think the Python interpreter will be baked into the Rust executable,
and thus any virtualenv you might be in at run-time might be for a
totally different version of Python.

The python part needs 'cma' pip package installed externally which is
where the virtualenv (and package directories in general) come into
play.