Make pyo3 an optional dependency through cargo feature.
This dependency causes some fuss. I don't want to boot it out entirely
because I still occasionally want to use the PyCMAES wrapping, but I
want the option to have pyo3 to be used when I ask for it.

Pyo3 is still a default feature to keep this backwards compatible with
how it was. But you can now disable default features to get "pycmaes"
feature disabled.