Move to `pkg-config` based finding of libcmaes and eigen3.
I had previously just hardcoded horrible homebrew paths and stuff.

On my machines, this pkg-config is working, although I had to add a
small hack for libcmaes, because the include path as given by pkg-config
does not seem like it is usable directly (it gives a path that's too
deep by one directory, so I had to put ../ on it).

This worked on MacOS and Linux. libcmaes tends to be too obscure to be
packaged by distributions, but a locally installed one will do (it
includes a pkg-config .pc file). Eigen3 worked with the pkg-config
without other fuss.